+++ RELEASE_NOTES.md
@@ -37,4 +37,5 @@ These release notes will only contain the current version. Just check out an old
 * `SystemKeyboardButtonBody` has been renamed to `KeyboardButtonBody`.
 * `SystemKeyboardButtonContent` has been renamed to `KeyboardButtonContent`.
 * `SystemKeyboardButtonShadow` has been moved into `KeyboardButtonBody`.
+* `SystemKeyboardButtonText` has been renamed to `KeyboardButtonText`.
 * `View.systemKeyboardButtonStyle(...)` has been renamed to `.keyboardButtonStyle(...)`.